Added endpoint for regions and also a method that searches region by its id (postal code)

+++ e-gov-backend/src/main/java/com/egovernment/egovbackend/domain/factory/region/RegionFactory.java
@@ -5,7 +5,7 @@ import com.egovernment.egovbackend.domain.entity.Region;
 public class RegionFactory implements RegionFactoryInterface {
 
     @Override
-    public Region createRegion(byte id, String englishRegionName
+    public Region createRegion(int id, String englishRegionName
             , String bulgarianRegionName){
         return Region.builder()
                 .id(id)
