+++ e-gov-backend/src/main/java/com/egovernment/egovbackend/domain/factory/region/RegionFactoryInterface.java
@@ -3,5 +3,5 @@ package com.egovernment.egovbackend.domain.factory.region;
 import com.egovernment.egovbackend.domain.entity.Region;
 
 public interface RegionFactoryInterface {
-    Region createRegion(byte id, String englishRegionName, String bulgarianRegionName);
+    Region createRegion(int id, String englishRegionName, String bulgarianRegionName);
 }
